little fixes

+++ gui/PlugUI.cpp
@@ -91,6 +91,7 @@ PlugUI::PlugUI (SynthParams &p)
     bpmDisplay->setColour (TextEditor::textColourId, Colours::black);
     bpmDisplay->setColour (TextEditor::backgroundColourId, Colour (0x00000000));
 
+
     //[UserPreSize]
     freq->setValue(params.freq.getUI());
     freq->setTextValueSuffix(String(" ") + params.freq.unit());
@@ -147,7 +148,6 @@ void PlugUI::resized()
     label2->setBounds (726, 8, 64, 16);
     bpmLabel->setBounds (8, 64, 56, 24);
     bpmDisplay->setBounds (56, 64, 150, 24);
-
     //[UserResized] Add your own custom resize handling here..
     //[/UserResized]
 }
@@ -178,7 +178,7 @@ void PlugUI::timerCallback()
 
 void PlugUI::updateBpmDisplay(const AudioPlayHead::CurrentPositionInfo &currentPos)
 {
-    lastBpmInfo = currentPos.bpm;
+    lastBpmInfo = static_cast<float>(currentPos.bpm);
 
     MemoryOutputStream bpmDisplayText;
